feat: build container for each version of the CLI (#24)
This commit adds a new feature to the project that builds a container
for every version. This is helpful for easily running the CLI without
having a local installation of Python, e.g. running the CLI in CI or on
Kubernetes.

With the current GitHub actions setup, a container is built on every
merge to main with the Git SHA as the tag as well as on every trigger of
the `release.yml` workflow, in which case the tag will be the version of
the release.

Images will be pushed to ghcr.io/fal-ai/fal.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>