fix: correct sorting of packages
Right now, all of the local packages are grouped with 3rd-party packages
because isort does not know that everything under
projects/fal is local. This is probably because
installing the root of the repository does not actually bring any
isolate_controller code with it, and because isort runs from the
perspective of the repository root, the tool understands that all
isolate_controller code is foreign.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>