fix(cli): use conventional __main__.py
This allows people to do run fal like so `python -m fal`.
I don't think anyone was using `python -m fal.cli`, but
have to clarify that this patch makes it no longer work,
which could be considered a breaking change.

Also it is common for people to expect a `__main__.py` to
be present, so this should be useful for anyone lurking around.