fix(auth): always try to open browser but also show url and code
The prompt is unnecessary, we should just try opening a browser automatically.

Currently the URL tells me to check the code in CLI, which is not shown to me
if `webbrowser.get()` was able to find a browser. Also then `open_new_tab` is
not really guaranteed to work, so we are better off just always showing the URL
as well.