feat: wrap tracebacks in JSON for easy detection
Right now, multi-line tracebacks are spread out and ingested
independently into the underlying logging system. This makes it hard to
retrieve the entirety of a traceback without some hacky heuristics. To
enable easy searching and retrieval of tracebacks, we catch all uncaught
exceptions and log them as JSON, this way the whole traceback is in one
single JSON line, which can be extracted by looking at the `traceback`
key.

An alternative would be to print the traceback in `unicode_escape`
encoding and then to to decode the string during log retrieval. This
would remove the need for a JSON wrapping.

Signed-off-by: squat <lserven@gmail.com>