misc(container): add tests for container apps and remove build_env (#375)
* misc(container): add tests for container apps and remove build_env

We currently do not exercise container apps in our tests, the container
app seems to be an unused fixture. This commit introduces basic tests to
ensure that container apps are built and work as expected. Furthermore,
this commit introduces a check that container args actually work.
Finally, this commit removes the `build_env` keyword from the
`ContainerImage` class. Docker has no concept of specifying environment
variables at build time except by declaring build arguments. Having that
keyword is messy and opens the door to non-hermetic builds, which is
a recipe for leaks and reproducibility problems.

Signed-off-by: squat <lserven@gmail.com>

* fix isort

---------

Signed-off-by: squat <lserven@gmail.com>